+++ core/src/main/java/ai/timefold/solver/core/api/score/stream/tri/TriConstraintStream.java
@@ -1304,9 +1304,11 @@ public interface TriConstraintStream<A, B, C> extends ConstraintStream {
             return firstStream.concat(secondStream.ifNotExists(remapped, Joiners.equal()),
                     paddingFunctionB, paddingFunctionC);
         } else {
-            var secondStream = getConstraintFactory().forEach(otherClass)
-                    .ifNotExists(remapped, Joiners.equal());
-            return firstStream.concat(secondStream, paddingFunctionB, paddingFunctionC);
+            throw new IllegalStateException("""
+                    Impossible state: the %s class (%s) does not extend %s.
+                    %s are not expected to be implemented by the user.
+                    """.formatted(ConstraintStream.class.getSimpleName(), this.getClass().getSimpleName(),
+                    AbstractConstraintStream.class.getSimpleName(), ConstraintStream.class.getSimpleName()));
         }
     }
