+++ core/src/test/java/ai/timefold/solver/core/impl/score/stream/bavet/tri/BavetTriConstraintStreamPrecomputeTest.java
@@ -1,0 +1,13 @@
+package ai.timefold.solver.core.impl.score.stream.bavet.tri;
+
+import ai.timefold.solver.core.impl.score.constraint.ConstraintMatchPolicy;
+import ai.timefold.solver.core.impl.score.stream.bavet.BavetConstraintStreamImplSupport;
+import ai.timefold.solver.core.impl.score.stream.common.tri.AbstractTriConstraintStreamPrecomputeTest;
+
+final class BavetTriConstraintStreamPrecomputeTest extends AbstractTriConstraintStreamPrecomputeTest {
+
+    public BavetTriConstraintStreamPrecomputeTest(ConstraintMatchPolicy constraintMatchPolicy) {
+        super(new BavetConstraintStreamImplSupport(constraintMatchPolicy));
+    }
+
+}
