+++ core/src/test/java/ai/timefold/solver/core/impl/score/stream/common/uni/AbstractUniConstraintStreamTest.java
@@ -37,9 +37,6 @@ import ai.timefold.solver.core.api.score.stream.Constraint;
 import ai.timefold.solver.core.api.score.stream.ConstraintCollectors;
 import ai.timefold.solver.core.api.score.stream.ConstraintProvider;
 import ai.timefold.solver.core.api.score.stream.DefaultConstraintJustification;
-import ai.timefold.solver.core.api.score.stream.Joiners;
-import ai.timefold.solver.core.api.score.stream.PrecomputeFactory;
-import ai.timefold.solver.core.api.score.stream.uni.UniConstraintStream;
 import ai.timefold.solver.core.impl.domain.solution.descriptor.SolutionDescriptor;
 import ai.timefold.solver.core.impl.score.director.InnerScoreDirector;
 import ai.timefold.solver.core.impl.score.stream.common.AbstractConstraintStreamTest;
@@ -75,7 +72,6 @@ import ai.timefold.solver.core.testdomain.unassignedvar.TestdataAllowsUnassigned
 import ai.timefold.solver.core.testdomain.unassignedvar.TestdataAllowsUnassignedSolution;
 
 import org.junit.jupiter.api.TestTemplate;
-import org.mockito.Mockito;
 
 public abstract class AbstractUniConstraintStreamTest
         extends AbstractConstraintStreamTest
@@ -3816,247 +3812,6 @@ public abstract class AbstractUniConstraintStreamTest
         assertScore(scoreDirector,
                 assertMatch(entityWithNull),
                 assertMatch(entityWithValue));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_filter_0_changed() {
-        var solution = TestdataLavishSolution.generateSolution();
-        var entityGroup = new TestdataLavishEntityGroup("MyEntityGroup");
-        solution.getEntityGroupList().add(entityGroup);
-        var entity1 = Mockito.spy(new TestdataLavishEntity("MyEntity 1", entityGroup, solution.getFirstValue()));
-        solution.getEntityList().add(entity1);
-        var entity2 = new TestdataLavishEntity("MyEntity 2", entityGroup, solution.getFirstValue());
-        solution.getEntityList().add(entity2);
-        var entity3 = new TestdataLavishEntity("MyEntity 3", solution.getFirstEntityGroup(),
-                solution.getFirstValue());
-        solution.getEntityList().add(entity3);
-
-        var scoreDirector =
-                buildScoreDirector(factory -> factory.precompute(data -> data.forEachUnfiltered(TestdataLavishEntity.class)
-                        .filter(entity -> entity.getEntityGroup() == entityGroup))
-                        .penalize(SimpleScore.ONE)
-                        .asConstraint(TEST_CONSTRAINT_NAME));
-
-        // From scratch
-        Mockito.reset(entity1);
-        scoreDirector.setWorkingSolution(solution);
-        assertScore(scoreDirector,
-                assertMatch(entity1),
-                assertMatch(entity2));
-        Mockito.verify(entity1, Mockito.atLeastOnce()).getEntityGroup();
-
-        // Incrementally update a variable
-        Mockito.reset(entity1);
-        scoreDirector.beforeVariableChanged(entity1, "value");
-        entity1.setValue(new TestdataLavishValue());
-        scoreDirector.afterVariableChanged(entity1, "value");
-        assertScore(scoreDirector,
-                assertMatch(entity1),
-                assertMatch(entity2));
-        Mockito.verify(entity1, Mockito.never()).getEntityGroup();
-
-        // Incrementally update a fact
-        scoreDirector.beforeProblemPropertyChanged(entity3);
-        entity3.setEntityGroup(entityGroup);
-        scoreDirector.afterProblemPropertyChanged(entity3);
-        assertScore(scoreDirector,
-                assertMatch(entity1),
-                assertMatch(entity2),
-                assertMatch(entity3));
-
-        // Remove entity
-        scoreDirector.beforeEntityRemoved(entity3);
-        solution.getEntityList().remove(entity3);
-        scoreDirector.afterEntityRemoved(entity3);
-        assertScore(scoreDirector,
-                assertMatch(entity1),
-                assertMatch(entity2));
-
-        // Add it back again, to make sure it was properly removed before
-        scoreDirector.beforeEntityAdded(entity3);
-        solution.getEntityList().add(entity3);
-        scoreDirector.afterEntityAdded(entity3);
-        assertScore(scoreDirector,
-                assertMatch(entity1),
-                assertMatch(entity2),
-                assertMatch(entity3));
-    }
-
-    private <T> void assertPrecompute(TestdataLavishSolution solution,
-            List<T> expectedValues,
-            Function<PrecomputeFactory, UniConstraintStream<T>> entityStreamSupplier) {
-        var scoreDirector =
-                buildScoreDirector(factory -> factory.precompute(entityStreamSupplier)
-                        .ifExists(TestdataLavishEntity.class)
-                        .penalize(SimpleScore.ONE)
-                        .asConstraint(TEST_CONSTRAINT_NAME));
-
-        // From scratch
-        scoreDirector.setWorkingSolution(solution);
-        assertScore(scoreDirector);
-
-        for (var entity : solution.getEntityList()) {
-            scoreDirector.beforeVariableChanged(entity, "value");
-            entity.setValue(solution.getFirstValue());
-            scoreDirector.afterVariableChanged(entity, "value");
-        }
-
-        assertScore(scoreDirector, expectedValues.stream()
-                .map(AbstractConstraintStreamTest::assertMatch)
-                .toArray(AssertableMatch[]::new));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_ifExists() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityWithGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .ifExists(TestdataLavishEntityGroup.class, Joiners.equal(
-                                TestdataLavishEntity::getEntityGroup, Function.identity())));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_ifNotExists() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityWithoutGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .ifNotExists(TestdataLavishEntityGroup.class, Joiners.equal(
-                                TestdataLavishEntity::getEntityGroup, Function.identity())));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_groupBy() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .filter(entity -> entity.getEntityGroup() != null)
-                        .groupBy(TestdataLavishEntity::getEntityGroup));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_flattenLast() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityWithoutGroup, entityWithGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .groupBy(ConstraintCollectors.toList())
-                        .flattenLast(entityList -> entityList));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_map() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup1 = new TestdataLavishEntity();
-        var entityWithGroup2 = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup1.setEntityGroup(entityGroup);
-        entityWithGroup2.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup1, entityWithGroup2));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityGroup, entityGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .filter(entity -> entity.getEntityGroup() != null)
-                        .map(TestdataLavishEntity::getEntityGroup));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_concat() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityWithoutGroup, entityWithGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .filter(entity -> entity.getEntityGroup() == null)
-                        .concat(pf.forEachUnfiltered(TestdataLavishEntity.class)
-                                .filter(entity -> entity.getEntityGroup() != null)));
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_distinct() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup1 = new TestdataLavishEntity();
-        var entityWithGroup2 = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup1.setEntityGroup(entityGroup);
-        entityWithGroup2.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup1, entityWithGroup2));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .filter(entity -> entity.getEntityGroup() != null)
-                        .map(TestdataLavishEntity::getEntityGroup)
-                        .distinct());
-    }
-
-    @Override
-    @TestTemplate
-    public void precompute_complement() {
-        var solution = TestdataLavishSolution.generateEmptySolution();
-        var entityWithoutGroup = new TestdataLavishEntity();
-        var entityWithGroup1 = new TestdataLavishEntity();
-        var entityWithGroup2 = new TestdataLavishEntity();
-        var entityGroup = new TestdataLavishEntityGroup();
-        entityWithGroup1.setEntityGroup(entityGroup);
-        entityWithGroup2.setEntityGroup(entityGroup);
-        solution.getEntityList().addAll(List.of(entityWithoutGroup, entityWithGroup1, entityWithGroup2));
-        solution.getEntityGroupList().add(entityGroup);
-        solution.getValueList().add(new TestdataLavishValue());
-
-        assertPrecompute(solution, List.of(entityWithGroup1, entityWithGroup2, entityWithoutGroup),
-                pf -> pf.forEachUnfiltered(TestdataLavishEntity.class)
-                        .filter(entity -> entity.getEntityGroup() != null)
-                        .complement(TestdataLavishEntity.class));
     }
 
     @TestTemplate
